+++ test/internal/db/schema.rb
@@ -40,6 +40,7 @@ ActiveRecord::Schema.define do
     t.string :check_type
     t.text :message
     t.datetime :last_run_at
+    t.datetime :paused_at
     t.timestamps null: false
   end